backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.5
httpx==0.25.1
python-dotenv==1.0.0
websockets==12.0